Implementation of Post View functionality.
