+++ frontend/src/components/posts/PostsPage.js
@@ -1,8 +1,30 @@
-import React from "react";
+import React, {useEffect, useState} from "react";
 import PostsApi from "../../api/PostsApi";
 import PostForm from "./PostForm";
+import ShowPost from "./ShowPost";
+import TrendingPost from "./TrendingPost";
 function PostsPage() {
     // call a REST API to show all the post .Depending on the no of post recieved ,populate the card div
+
+    const [information, setInformation] = useState([]);
+    const viewPosts = () => {
+
+        PostsApi.getAllPosts()
+            .then(response => {
+
+                setInformation(response.data);
+            })
+
+    }
+
+
+
+    useEffect(() => {
+        viewPosts();
+
+    }, []);
+
+
     const createPost = (postData) => {
 
             PostsApi.createPost(postData)
@@ -12,6 +34,9 @@ function PostsPage() {
     return (
         <div>
              <PostForm onSubmit={createPost}/>
+             <TrendingPost information={information}/>
+
+
         </div>
     );
 }
